Step5 - 로또(수동) PR 요청드립니다. (#920)
* docs: 5단계 요구사항 정리

* feat: 수동으로 구매할 로또 입력하는 기능

* feat: 수동으로 입력받은 숫자를 로또로 만드는 기능

* feat: 자동,수동 로또를 병합하는 기능

* feat: 수동 로또 출력 기능 추가

* feat: controller에 로또 수동 구매 로직 추가

* feat: LottoUtil에 계산 기능 추가

* feat: LottoUtil 테스트 추가

* docs: 요구사항 구현 체크

* refactor: 주석제거

* refactor: LottoGenerateStrategy 추가, 메소드 분리

* refactor: 출력 오류 수정

+++ src/main/java/step3/domain/Rewards.java
@@ -3,7 +3,7 @@ package step3.domain;
 import java.util.List;
 
 public class Rewards {
-    private static final long LOTTO_PRICE = 1000L;
+    private static final int LOTTO_PRICE = 1000;
     private final List<Rank> rankList;
     private final long totalPrize;
